# demo-api/openapi.yaml
openapi: 3.0.0
info:
  title: RICO Security Testing Playground
  description: |
    Intentionally vulnerable API for testing RICO's security scanning capabilities.
    
    ⚠️ WARNING: This API contains intentional security vulnerabilities.
    DO NOT deploy this in production environments.
    
    ## Vulnerability Categories
    - SQL Injection (error-based, boolean-blind, time-based)
    - IDOR (Insecure Direct Object Reference)
    - Missing Authentication
    - SSRF (Server-Side Request Forgery)
    - Path Traversal
    - Command Injection
    - JWT Tampering (signature bypass, alg=none)
    - Business Logic Flaws (negative transfers, role injection)
    
    ## Testing Credentials
    - Regular User: alice / password123
    - Regular User: bob / password456
    - Admin User: admin / admin123
  version: 1.0.0
  contact:
    name: RICO Security
    url: https://github.com/yourusername/rico
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /users/{user_id}/secure:
    get:
      tags: [Users]
      summary: Get user by ID (SECURE)
      description: |
        🟢 SECURE: Proper authorization - users can only access their own data.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User data
        '403':
          description: Forbidden
        '404':
          description: User not found
...
  /business/transfer/vulnerable:
    post:
      tags: [Business Logic]
      summary: Money transfer (VULNERABLE - Business Logic Flaw)
      description: |
        🔴 VULNERABLE: Multiple business logic flaws.
        
        Vulnerabilities:
        - Allows negative amounts (money creation)
        - No ownership verification
        - No balance check
        
        Test: Transfer -100 to create money!
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from_account:
                  type: integer
                to_account:
                  type: integer
                amount:
                  type: number
      responses:
        '200':
          description: Transfer successful
  /business/transfer/secure:
    post:
      tags: [Business Logic]
      summary: Money transfer (SECURE)
      description: |
        🟢 SECURE: Proper validation and authorization.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                from_account:
                  type: integer
                to_account:
                  type: integer
                amount:
                  type: number
      responses:
        '200':
          description: Transfer successful
        '400':
          description: Invalid request
        '403':
          description: Forbidden
  /business/orders/vulnerable:
    post:
      tags: [Business Logic]
      summary: Create order (VULNERABLE - Role Injection)
      description: |
        🔴 VULNERABLE: Trusts user-provided role in nested JSON.
        
        Test payload:
        ```json
        {
          "user": {"id": 1, "role": "admin"},
          "items": [{"product_id": 101, "quantity": 2}]
        }
        ```
        User can set their own role to get admin discount!
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Order created
  /admin/users/vulnerable:
    get:
      tags: [Admin]
      summary: List users (VULNERABLE - JWT Tampering)
      description: |
        🔴 VULNERABLE: JWT signature not validated.
        
        JWT Tampering test:
        1. Login as regular user
        2. Modify JWT: change "role": "user" to "role": "admin"
        3. Use modified token (signature not checked!)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list
        '401':
          description: Unauthorized
  /admin/users/secure:
    get:
      tags: [Admin]
      summary: List users (SECURE)
      description: |
        🟢 SECURE: Proper JWT validation and role check.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /admin/promote/{user_id}/vulnerable:
    post:
      tags: [Admin]
      summary: Promote user (VULNERABLE - No Auth)
      description: |
        🔴 VULNERABLE: No authentication required!
        
        JWT alg=none bypass test:
        1. Create JWT with alg=none
        2. No signature required
        3. Access endpoint
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User promoted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT